requirements.txt
numpy>=1.21
Pillow>=8.0.0
dataset==1.1.2
facenet_pytorch==0.1.0
Flask==1.1.1
flask_restplus==0.13.0
joblib>=0.13.2
matplotlib>=3.5
seaborn>=0.9.0
scikit_learn>=0.21.3
torch>=2.0.0
torchvision>=0.15
Werkzeug>=0.15.2
opencv-python>=4.5
